pyproject.toml
[tool.poetry]
name = "mozyq"
version = "0.0.2"
description = ""
authors = ["Gilad Kutiel <gilad.kutiel@gmail.com>"]
readme = "README.md"

[tool.poetry.group.dev.dependencies]
pyright = "^1.1.375"
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
cattrs = "^23.2.3"
opencv-python = "^4.10.0.84"
scipy = "^1.14.0"
torchaudio = "^2.4.0"
torchvision = "^0.19.0"
tqdm = "^4.66.5"
typer = "^0.12.3"

[tool.poetry.dev-dependencies]
pre-commit = "^3.8.0"

[tool.poetry.scripts]
mzq = "mozyq.app:main"
